RendererDRMPRIMEGLES: Fix leak of EGL fences
When running on wayland in a desktop window, kodi will crash with
'Too many open files' after a minute of so of video playback.

I've tracked it down the EGL fences. Render gets called more often than ReleaseBuffer
(I'm seeing two Render calls per ReleaseBuffer call).

As we allocate a fence in Render and free in ReleaseBuffer this leaks.

It's safe to call DestroyFence on a non-created fence so use that to stop the leak